AL-BackEnd/requirements.txt
# pip install -r requirements.txt
flask
pycryptodome
requests
beautifulsoup4
flask-cors
python-dotenv